api/requirements.txt
# api/requirements.txt
requests
google-generativeai
markdown2